Base version with typescript. Refactoring required.
