--- app/src/main.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Page from './page/page.jsx';
-import Data from '../LinksWithStars.json';
-import './style.less'
-import 'css!normalize.css'
-
-ReactDOM.render(
-    <Page categories={Data}/>,
-    document.getElementById('root')
-);
