+++ app/src/main.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+import {Page} from './page/page';
+import './style.less'
+import 'css!normalize.css'
+
+const Data = require('../LinksWithStars.json');
+
+ReactDOM.render(
+    <Page categories={Data}/>,
+    document.getElementById('root')
+);
+
