+++ package.json
@@ -8,21 +8,20 @@
     "lodash": "^4.15.0",
     "normalize.css": "^4.2.0",
     "react": "^15.3.1",
-    "react-dom": "^15.3.1"
+    "react-dom": "^15.3.1",
+    "react-router": "^3.0.0"
   },
   "devDependencies": {
+    "@types/history": "^2.0.39",
     "@types/lodash": "^4.14.38",
     "@types/node": "^6.0.46",
     "@types/node-sass": "^3.10.30",
     "@types/react": "^0.14.47",
     "@types/react-dom": "^0.14.18",
+    "@types/react-router": "^2.0.38",
     "@types/request": "^0.0.33",
     "@types/rss": "^0.0.27",
     "autoprefixer": "^6.4.0",
-    "babel-core": "^6.14.0",
-    "babel-loader": "^6.2.5",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
     "css-loader": "^0.24.0",
     "file-loader": "^0.9.0",
     "front-matter": "^2.1.0",
@@ -41,6 +40,7 @@
     "rss": "^1.2.1",
     "sitemap": "^1.8.1",
     "style-loader": "^0.13.1",
+    "ts-loader": "^1.1.0",
     "ts-node": "^1.7.0",
     "typescript": "^2.0.8",
     "url-loader": "^0.5.7",
@@ -52,12 +52,12 @@
     "prestart": "yarn install",
     "start": "webpack-dev-server",
     "pretest": "yarn install",
-    "test": "ts-node ./app/LinkChecker.ts",
-    "stars": "ts-node ./app/Stars.ts",
-    "sitemap": "ts-node ./app/Sitemap.ts",
-    "rss": "ts-node ./app/rss/RssGenerator.ts",
-    "pages": "ts-node ./app/rss/PagesGenerator.ts",
-    "readme": "ts-node ./app/ReadmeGenerator.ts",
+    "test": "ts-node -n --compilerOptions '{\"target\":\"es6\",\"module\":\"commonjs\"}' ./app/LinkChecker.ts",
+    "stars": "ts-node -n --compilerOptions '{\"target\":\"es6\",\"module\":\"commonjs\"}' ./app/Stars.ts",
+    "sitemap": "ts-node -n --compilerOptions '{\"target\":\"es6\",\"module\":\"commonjs\"}' ./app/Sitemap.ts",
+    "rss": "ts-node -n --compilerOptions '{\"target\":\"es6\",\"module\":\"commonjs\"}' ./app/rss/RssGenerator.ts",
+    "pages": "ts-node -n --compilerOptions '{\"target\":\"es6\",\"module\":\"commonjs\"}' ./app/rss/PagesGenerator.ts",
+    "readme": "ts-node -n --compilerOptions '{\"target\":\"es6\",\"module\":\"commonjs\"}' ./app/ReadmeGenerator.ts",
     "pack": "export NODE_ENV=production; export SOURCE_MAP=source-map; webpack",
     "ci": "npm run stars && npm run pack && npm run rss && npm run pages && npm run sitemap",
     "build": "npm run stars && npm run readme && npm run pack && npm run rss"
